Add new files and update existing files
