--- old/app/routes/_index/components/pages/index.ts
@@ -1,3 +1,0 @@
-export * from './AboutPage'
-export * from './ContactPage'
-export * from './HeroPage'
