+++ old/app/routes/api.contact.tsx
@@ -3,7 +3,7 @@ import { getFieldsetConstraint, parse } from '@conform-to/zod'
 import { Link, useFetcher } from '@remix-run/react'
 import { json, type ActionFunctionArgs } from '@vercel/remix'
 import { z } from 'zod'
-import PrivacyPolicyDialog from '~/components/PrivacyPolicyDialog'
+import PrivacyPolicyDialog from '../../../src/components/PrivacyPolicyDialog'
 import { Button, Checkbox, HStack, Input, Label, Stack, Textarea } from '~/components/ui'
 import { useLocale } from '~/features/i18n/hooks/useLocale'
 import { sendEmail } from '~/services/sendEmail'
