+++ old/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "include": ["*.js", "**/*.ts", "**/*.tsx"],
+  "include": ["*.js", "**/*.ts", "**/*.tsx", "../src/components/PrivacyPolicyDialog.tsx"],
   "compilerOptions": {
     "lib": ["DOM", "DOM.Iterable", "ES2022"],
     "types": ["@types/node"],
