+++ prettier.config.js
@@ -1,0 +1,8 @@
+export default {
+  semi: false,
+  singleQuote: true,
+  trailingComma: 'all',
+  printWidth: 120,
+  tabWidth: 2,
+  plugins: ['prettier-plugin-astro', 'prettier-plugin-tailwindcss', 'prettier-plugin-organize-imports'],
+}
