+++ src/components/CoverPage.astro
@@ -1,0 +1,15 @@
+---
+const { id, bgImage } = Astro.props
+---
+
+<div
+  id={id}
+  class="relative flex h-screen snap-start snap-always items-center justify-center bg-black/40 bg-cover bg-center bg-blend-overlay"
+  style={{
+    backgroundImage: `url(${bgImage})`,
+  }}
+>
+  <div class="mx-auto w-full px-4 text-center sm:w-2/3 sm:px-0">
+    <slot />
+  </div>
+</div>
