+++ src/components/Navigation.astro
@@ -1,0 +1,42 @@
+---
+import HStack from '~/components/ui/HStack.astro'
+import Spacer from '~/components/ui/Spacer.astro'
+
+const t = (key: string, fallback: string) => {
+  const lang = 'ja'
+  const translations: Record<string, Record<string, string>> = {
+    'nav.techtalk': {
+      ja: 'TechTalk',
+      en: 'TechTalk',
+    },
+    'nav.about': {
+      ja: '会社概要',
+      en: 'About',
+    },
+    'nav.contact': {
+      ja: 'お問い合わせ',
+      en: 'Contact',
+    },
+  }
+
+  return translations[key][lang] || fallback
+}
+---
+
+<HStack class="fixed z-10 w-full px-4 py-2 font-bold text-white sm:pr-8">
+  <a href="#hero" class="cursor-pointer hover:text-indigo-500">
+    {t('nav.techtalk', 'TechTalk')}
+  </a>
+
+  <Spacer />
+  <HStack>
+    <a href="#about" class="cursor-pointer hover:text-indigo-500">
+      {t('nav.about', '会社概要')}
+    </a>
+    <a href="#contact" class="cursor-pointer hover:text-indigo-500">
+      {t('nav.contact', 'お問い合わせ')}
+    </a>
+
+    <!-- <LanguageSwitcher /> -->
+  </HStack>
+</HStack>
