+++ src/components/pages/index.ts
@@ -1,0 +1,3 @@
+export * from "./AboutPage.astro";
+export * from "./ContactPage";
+export * from "./HeroPage";
