+++ src/components/ui/Spacer.astro
@@ -1,0 +1,5 @@
+---
+
+---
+
+<div class="flex-1"></div>
