+++ src/components/ui/heading.astro
@@ -1,0 +1,29 @@
+---
+import { cva } from 'class-variance-authority'
+import { cn } from '~/libs/utils'
+
+const headingVariants = cva('font-sans font-black', {
+  variants: {
+    size: {
+      sm: 'text-sm',
+      md: 'text-md',
+      lg: 'text-lg',
+      xl: 'text-xl',
+      '2xl': 'text-2xl',
+      '4xl': 'text-4xl',
+      '5xl': 'text-5xl',
+      '6xl': 'text-6xl',
+      '8xl': 'text-8xl',
+    },
+  },
+  defaultVariants: {
+    size: '2xl',
+  },
+})
+
+const { As, size } = Astro.props.As ?? 'h2'
+---
+
+<As class={cn(headingVariants({ size }))}>
+  <slot />
+</As>
