+++ src/components/ui/index.ts
@@ -4,7 +4,6 @@ export * from "./card";
 export * from "./checkbox";
 export * from "./dialog";
 export * from "./dropdown-menu";
-export * from "./heading";
 export * from "./input";
 export * from "./label";
 export * from "./popover";
