+++ src/layouts/Layout.astro
@@ -1,19 +1,20 @@
 ---
 import '~/styles/globals.css'
+import Navigation from '~/components/Navigation.astro'
 ---
 
 <html lang="ja">
-	<head>
-		<meta charset="utf-8" />
-		<meta name="description" content="株式会社TechTalkの公式サイトです。" />
-		<link rel="icon" type="image/svg+xml" href="/favicon.svg" />
-		<meta name="viewport" content="width=device-width,initial-scale=1" />
-		<meta name="generator" content={Astro.generator} />
-		<title>株式会社TechTalk</title>
-	</head>
+  <head>
+    <meta charset="utf-8" />
+    <meta name="description" content="株式会社TechTalkの公式サイトです。" />
+    <link rel="icon" type="image/svg+xml" href="/favicon.svg" />
+    <meta name="viewport" content="width=device-width,initial-scale=1" />
+    <meta name="generator" content={Astro.generator} />
+    <title>株式会社TechTalk</title>
+  </head>
 
-  <body class="relative h-screen text-white bg-black">
-    <div>navigation</div>
+  <body class="relative h-screen bg-black text-white">
+    <Navigation />
     <slot />
-	</body>
+  </body>
 </html>
