+++ src/pages/index.astro
@@ -1,13 +1,16 @@
 ---
+import { getRelativeLocaleUrl } from 'astro:i18n'
 import Layout from '~/layouts/Layout.astro'
-import { LanguageSwitcher} from '~/features/i18n/components/LanguageSwitcher'
+import LanguageSwitcher from '~/features/i18n/components/LanguageSwitcher.astro'
 ---
-<Layout>
-	<main class="h-screen snap-y snap-mandatory overflow-auto">
-		<LanguageSwitcher client:load client:only />
 
-		<div>ヒーロー page</div>
-		<div>about page</div>
-		<div>contact page</div>
-	</main>
+<Layout>
+  <main class="h-screen snap-y snap-mandatory overflow-auto">
+    <a href={getRelativeLocaleUrl('ja', '/')}>ja</a>
+    <a href={getRelativeLocaleUrl('en', '/')}>en</a>
+
+    <div>TechTalk</div>
+    <div>会社概要</div>
+    <div>お問い合わせ</div>
+  </main>
 </Layout>
